[src/{sqrt_ui.c,sub_ui.c,ui_div.c,ui_sub.c}] Changed cnt type from unsigned long to int (signed) due to the MPFR_SET_EXP requirement. Otherwise the code was correct. Now, 160 tests pass and 9 tests fail on Linux/x86_64.
[From SVN r8938 (branches/set_exp)]

+++ src/sub_ui.c
@@ -31,7 +31,7 @@ mpfr_sub_ui (mpfr_ptr y, mpfr_srcptr x, unsigned long int u, mpfr_rnd_t rnd_mode
     {
       mpfr_t uu;
       mp_limb_t up[1];
-      unsigned long cnt;
+      int cnt;
       int inex;
 
       MPFR_SAVE_EXPO_DECL (expo);
